refactor: extract the shared JSONC settings reader out of lint and load
lint/parse.ts re-declared settings/load.ts's entire JSONC reading layer
(KNOWN_EVENT_NAMES, getProperty, the require* helpers, and the read/parse/
hooks-object prologue), with only the matcher read actually differing in
strictness. Pulled the shared parts into settings/jsonc.ts so the two
readers cannot drift apart from each other by accident, and collapsed
src/cli.ts's three copies of the settings-discovery prologue (existence
check + discoverSources) into one resolveSettingsSources helper.

Adds a drift test in tests/lint.test.ts that walks every settings fixture
the strict loader accepts and asserts readHookCommands agrees with
loadSourceHooks on every (event, command, args, line) tuple.

Closes #49
